adding support for ODBC
Now it's possible two choose between the ODBC version ('vertica+odbc://...') or the vertica-python version ('vertica+vertica_python://...')
