+++ requirements.txt
@@ -1,3 +1,4 @@
-psycopg2 >= 2.6.2
 sqlalchemy >= 1.1.5
+psycopg2 >= 2.6.2
+pyodbc >= 3.0.10
 vertica-python >= 0.6.13
