+++ setup.py
@@ -3,12 +3,12 @@ from setuptools import setup
 with open("README.rst", "r") as f:
     description = f.read()
 
-VERSION = '0.2.0'
+VERSION = '0.2.1'
 
 setup(
     name='sqlalchemy-vertica',
     version=VERSION,
-    description='Vertica dialect for sqlalchemy using vertica_python',
+    description='Vertica dialect for sqlalchemy',
     long_description=description,
     license="MIT",
     url='https://github.com/startappdev/sqlalchemy-vertica',
@@ -19,12 +19,14 @@ setup(
         'sqlalchemy_vertica',
     ],
     install_requires=[
-        'psycopg2 >= 2.6.2',
         'sqlalchemy >= 1.1.5',
+        'psycopg2 >= 2.6.2',
+        'pyodbc >= 3.0.10',
         'vertica-python >= 0.6.13',
     ],
     entry_points="""
     [sqlalchemy.dialects]
-    vertica.vertica_python = sqlalchemy_vertica.base:VerticaDialect
+    vertica.pyodbc = sqlalchemy_vertica.odbc:VerticaDialect
+    vertica.vertica_python = sqlalchemy_vertica.vp:VerticaDialect
     """,
 )
