+++ sqlalchemy_vertica/odbc.py
@@ -1,0 +1,7 @@
+from sqlalchemy.connectors.pyodbc import PyODBCConnector
+
+from base import VerticaDialect as BaseVerticaDialect
+
+
+class VerticaDialect(PyODBCConnector, BaseVerticaDialect):
+    driver = 'pyodbc'
